EbAppConfig: Rewrite find_token to fix assertion
The way find_token was written would lead to it unconditionally
using the next argument following the token. In cases where the
token is the last argument, this could lead to it passing NULL
to strcpy_s which is invalid, triggering an assertion on Windows
debug builds.

This was not reproducible on other platforms as the replacement
strcpy_s implementations does not seem to follow the Windows
implementation for that and only returns an error for these
cases, which is ignored here.